Separate data sets for local and production setups.
