+++ components/pages/Homepage/index.tsx
@@ -1,6 +1,6 @@
 import styles from "../../../styles/Home.module.css";
 
-import data from "../../../data";
+import data, { EventFromJSON } from "../../../data";
 import { EventCard } from "./EventCard";
 
 export const Frontpage = () => {
@@ -8,7 +8,7 @@ export const Frontpage = () => {
     <>
       <h1 className={styles.title}>Tapahtumat</h1>
       <div className={styles.events}>
-        {data.map(item => (
+        {data.map((item: EventFromJSON) => (
           <EventCard key={item.id} event={item} />
         ))}
       </div>
