+++ data/production/index.js
@@ -1,0 +1,5 @@
+import events0 from "./2021/06.json";
+import events1 from "./2021/07.json";
+
+const data = [...events0, ...events1];
+export default data;
